==> code/00b_crosswalk_county_division.ipynb <==
# import packages
import numpy as np
import pandas as pd
from pathlib import Path
import requests

# Import data
home_dir = Path.home()
work_dir = (home_dir / 'Desktop' / 'GitHub' / 'election_inflation_analysis')
data = (work_dir / 'data')
output = work_dir / 'output'
clean_data = data / 'clean'
raw_data = (data / 'raw')
code = Path.cwd()



# State Abbreviations

url = "https://www2.census.gov/geo/docs/reference/state.txt"
response = requests.get(url)
if response.status_code == 200:
    with open("state.txt", "wb") as file:
        file.write(response.content)
    print("File downloaded successfully!")
else:
    print(f"Failed to download the file. Status code: {response.status_code}")

columns = ['STATE', 'STUSAB', 'STATE_NAME', 'STATENS']
states = pd.read_csv("state.txt", sep="|", names=columns, skiprows=1)

# Handle data
states.columns = states.columns.str.lower()

states.rename(columns = {'stusab': 'state_abbr'}, inplace= True)
keep = ['state_abbr', 'state_name']
states = states[keep]
states.rename(columns = {'state_name': 'state'}, inplace= True)
states = states

states = states.map(lambda x: x.lower() if isinstance(x, str) else x)

# Save the DataFrame as a CSV
states_output = clean_data / "state_abbr.csv"
states.to_csv(states_output, index=False)

# State and Census Divisions

# Crosswalk between States and Divisions
data = {
    "divisions": [
        "New England", "New England", "New England", "New England", "New England", "New England",
        "Middle Atlantic", "Middle Atlantic", "Middle Atlantic",
        "South Atlantic", "South Atlantic", "South Atlantic", "South Atlantic", "South Atlantic",
        "South Atlantic", "South Atlantic", "South Atlantic", "South Atlantic",
        "East South Central", "East South Central", "East South Central", "East South Central",
        "West South Central", "West South Central", "West South Central", "West South Central",
        "East North Central", "East North Central", "East North Central", "East North Central", "East North Central",
        "West North Central", "West North Central", "West North Central", "West North Central", 
        "West North Central", "West North Central", "West North Central",
        "Mountain", "Mountain", "Mountain", "Mountain", "Mountain", "Mountain", "Mountain", "Mountain",
        "Pacific", "Pacific", "Pacific", "Pacific", "Pacific"
    ],
    "state": [
        "Connecticut", "Maine", "Massachusetts", "New Hampshire", "Rhode Island", "Vermont",
        "New Jersey", "New York", "Pennsylvania",
        "Delaware", "District of Columbia", "Florida", "Georgia", "Maryland", "North Carolina",
        "South Carolina", "Virginia", "West Virginia",
        "Alabama", "Kentucky", "Mississippi", "Tennessee",
        "Arkansas", "Louisiana", "Oklahoma", "Texas",
        "Illinois", "Indiana", "Michigan", "Ohio", "Wisconsin",
        "Iowa", "Kansas", "Minnesota", "Missouri", "Nebraska", "North Dakota", "South Dakota",
        "Arizona", "Colorado", "Idaho", "Montana", "Nevada", "New Mexico", "Utah", "Wyoming",
        "Alaska", "California", "Hawaii", "Oregon", "Washington"
    ]
}

# Create a DataFrame
df_division = pd.DataFrame(data)

# Save the DataFrame as a CSV
output_file_division = clean_data / "state_division_crosswalk.csv"
df_division.to_csv(output_file_division, index=False)


### MSA County Crosswalk

crosswalk = raw_data / 'county_msa_crosswalk_raw.xlsx'
msa_county = pd.read_excel(crosswalk, sheet_name='Jul. 2023 Crosswalk')
msa_county = msa_county.rename(columns=str.lower).map(lambda x: x.lower() if isinstance(x, str) else x)
msa_county.rename(columns = {'county title': 'county'}, inplace = True)

# shorten msa to merge with zillow
msa_county['msa_name'] = msa_county['msa title'].str.split(',').str[0]
msa_county['msa_first'] = msa_county['msa_name'].str.split('-').str[0]
msa_county['state'] = msa_county['msa title'].str.extract(r', (\w{2}) msa')
msa_county['msa_short'] = msa_county['msa_first'] + ', ' + msa_county['state']
msa_county['county'] = msa_county['county'].str.replace(' county', '', regex = False)
crosswalk_file = clean_data / 'msa_county_crosswalk.csv'
msa_county.to_csv(crosswalk_file)
